+++ .storybook/config.js
@@ -60,6 +60,7 @@ configure(() => {
   require('./../packages/bpk-component-code/stories');
   require('./../packages/bpk-component-definition-list/stories');
   require('./../packages/bpk-component-content-container/stories');
+  require('./../packages/bpk-component-datatable/stories');
   require('./../packages/bpk-component-datepicker/stories');
   require('./../packages/bpk-component-drawer/stories');
   require('./../packages/bpk-component-fieldset/stories');